+++ src/Microsoft.Diagnostics.Runtime/ClrInstanceField.cs
@@ -18,110 +18,10 @@ namespace Microsoft.Diagnostics.Runtime
     /// </summary>
     public sealed class ClrInstanceField : ClrField, IClrInstanceField
     {
-        private int _attributes = (int)FieldAttributes.ReservedMask;
-
-        private readonly IClrTypeHelpers _helpers;
-        private readonly FieldInfo _data;
-        private string? _name;
-        private ClrType? _type;
-
         internal ClrInstanceField(ClrType containingType, ClrType? type, IClrTypeHelpers helpers, in FieldInfo data)
+            : base(containingType, type, helpers, data)
         {
-            _helpers = helpers;
-            _data = data;
-            ContainingType = containingType;
-
-            _type = type;
-            if (_data.ElementType == ClrElementType.Class && _type != null)
-                _data.ElementType = _type.ElementType;
-
-            DebugOnlyLoadLazyValues();
         }
-
-        [Conditional("DEBUG")]
-        private void DebugOnlyLoadLazyValues()
-        {
-            InitData();
-        }
-
-        private void InitData()
-        {
-            if (_attributes != (int)FieldAttributes.ReservedMask)
-                return;
-
-            MetadataImport? import = ContainingType.Module.MetadataImport;
-            if (import is null || !_helpers.GetFieldMetadataInfo(import, Token, out FieldMetadataInfo info))
-            {
-                _attributes = 0;
-                return;
-            }
-
-            StringCaching options = ContainingType.Heap.Runtime.DataTarget.CacheOptions.CacheFieldNames;
-            if (info.Name != null)
-            {
-                if (options == StringCaching.Intern)
-                    info.Name = string.Intern(info.Name);
-
-                if (options != StringCaching.None)
-                    _name = info.Name;
-            }
-
-            if (_type is null)
-            {
-                SigParser sigParser = new(info.Signature, info.SignatureSize);
-                if (sigParser.GetCallingConvInfo(out int sigType) && sigType == SigParser.IMAGE_CEE_CS_CALLCONV_FIELD)
-                {
-                    sigParser.SkipCustomModifiers();
-                    _type = ContainingType.Heap.GetOrCreateTypeFromSignature(ContainingType.Module, sigParser, ContainingType.EnumerateGenericParameters(), Array.Empty<ClrGenericParameter>());
-                }
-            }
-
-            Interlocked.Exchange(ref _attributes, (int)info.Attributes);
-        }
-
-        public override FieldAttributes Attributes
-        {
-            get
-            {
-                InitData();
-                return (FieldAttributes)_attributes;
-            }
-        }
-
-        public override ClrElementType ElementType => _data.ElementType;
-        public override bool IsObjectReference => ElementType.IsObjectReference();
-        public override bool IsValueType => ElementType.IsValueType();
-        public override bool IsPrimitive => ElementType.IsPrimitive();
-
-        public override string? Name
-        {
-            get
-            {
-                if (_name != null)
-                    return _name;
-
-                InitData();
-                return _name;
-            }
-        }
-
-        public override ClrType? Type
-        {
-            get
-            {
-                if (_type != null)
-                    return _type;
-
-                InitData();
-                return _type;
-            }
-        }
-
-        public override int Token => _data.Token;
-
-        public override int Offset => _data.Offset;
-
-        public override ClrType ContainingType { get; }
 
         /// <summary>
         /// Reads the value of the field as an unmanaged struct or primitive type.
