+++ src/Microsoft.Diagnostics.Runtime/ClrStaticField.cs
@@ -1,13 +1,8 @@
 ﻿// Licensed to the .NET Foundation under one or more agreements.
 // The .NET Foundation licenses this file to you under the MIT license.
 
-using System;
-using System.Reflection;
-using System.Threading;
 using Microsoft.Diagnostics.Runtime.AbstractDac;
-using Microsoft.Diagnostics.Runtime.DacInterface;
 using Microsoft.Diagnostics.Runtime.Interfaces;
-using Microsoft.Diagnostics.Runtime.Utilities;
 using FieldInfo = Microsoft.Diagnostics.Runtime.AbstractDac.FieldInfo;
 
 namespace Microsoft.Diagnostics.Runtime
@@ -17,100 +12,9 @@ namespace Microsoft.Diagnostics.Runtime
     /// </summary>
     public sealed class ClrStaticField : ClrField, IClrStaticField
     {
-        private readonly IClrTypeHelpers _helpers;
-        private readonly FieldInfo _data;
-        private string? _name;
-        private ClrType? _type;
-        private int _attributes = (int)FieldAttributes.ReservedMask;
-
         internal ClrStaticField(ClrType containingType, ClrType? type, IClrTypeHelpers helpers, in FieldInfo data)
+            : base(containingType, type, helpers, data)
         {
-            ContainingType = containingType;
-            _helpers = helpers;
-            _data = data;
-            _type = type;
-        }
-
-        public override ClrElementType ElementType => _data.ElementType;
-
-        public override bool IsObjectReference => ElementType.IsObjectReference();
-
-        public override bool IsValueType => ElementType.IsValueType();
-
-        public override bool IsPrimitive => ElementType.IsPrimitive();
-
-        public override string? Name
-        {
-            get
-            {
-                if (_name != null)
-                    return _name;
-
-                InitData();
-                return _name;
-            }
-        }
-
-        public override ClrType? Type
-        {
-            get
-            {
-                if (_type != null)
-                    return _type;
-
-                InitData();
-                return _type;
-            }
-        }
-
-        public override int Token => _data.Token;
-
-        public override int Offset => _data.Offset;
-
-        public override ClrType ContainingType { get; }
-
-        public override FieldAttributes Attributes
-        {
-            get
-            {
-                InitData();
-                return (FieldAttributes)_attributes;
-            }
-        }
-
-        private void InitData()
-        {
-            if (_attributes != (int)FieldAttributes.ReservedMask)
-                return;
-
-            MetadataImport? import = ContainingType.Module.MetadataImport;
-            if (import is null || !_helpers.GetFieldMetadataInfo(import, Token, out FieldMetadataInfo info))
-            {
-                _attributes = 0;
-                return;
-            }
-
-            StringCaching options = ContainingType.Heap.Runtime.DataTarget.CacheOptions.CacheFieldNames;
-            if (info.Name != null)
-            {
-                if (options == StringCaching.Intern)
-                    info.Name = string.Intern(info.Name);
-
-                if (options != StringCaching.None)
-                    _name = info.Name;
-            }
-
-            if (_type is null)
-            {
-                SigParser sigParser = new(info.Signature, info.SignatureSize);
-                if (sigParser.GetCallingConvInfo(out int sigType) && sigType == SigParser.IMAGE_CEE_CS_CALLCONV_FIELD)
-                {
-                    sigParser.SkipCustomModifiers();
-                    _type = ContainingType.Heap.GetOrCreateTypeFromSignature(ContainingType.Module, sigParser, ContainingType.EnumerateGenericParameters(), Array.Empty<ClrGenericParameter>());
-                }
-            }
-
-            Interlocked.Exchange(ref _attributes, (int)info.Attributes);
         }
 
         /// <summary>
